pyproject.toml
[project]
name = "structure-it"
version = "0.1.0"
description = "Research sandbox for structuring data with LLM-driven pipelines"
readme = "README.md"
requires-python = ">=3.11"
license = {text = "Apache-2.0"}
authors = [
    {name = "Fred Bliss"}
]
keywords = [
    "llm",
    "data-pipeline",
    "structured-output",
    "gemini",
    "research"
]
classifiers = [
    "Intended Audience :: Developers",
    "License :: OSI Approved :: Apache Software License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
]

dependencies = [
    "google-genai>=1.16.0",
    "pydantic>=2.0.0",
    "pydantic-settings>=2.0.0",
    "duckdb>=1.0.0",
    "python-dotenv>=1.0.0",
    "markitdown>=0.0.1",
    "civic-scraper>=0.1.0",
    "tenacity>=8.0.0",
    "scrapy>=2.13.4",
]

[project.optional-dependencies]
processors = [
    "genai-processors>=1.1.0",
]
server = [
    "fastapi>=0.109.0",
    "uvicorn>=0.27.0",
    "python-multipart>=0.0.9",
]
dev = [
    "pytest>=8.0.0",
    "pytest-cov>=4.1.0",
    "pytest-asyncio>=0.23.0",
    "ruff>=0.3.0",
    "mypy>=1.9.0",
]

[project.urls]
Homepage = "https://github.com/fredbliss/structure-it"
Repository = "https://github.com/fredbliss/structure-it"

[build-system]
requires = ["setuptools>=68.0", "wheel"]
build-backend = "setuptools.build_meta"

[tool.setuptools.packages.find]
where = ["src"]

[tool.pytest.ini_options]
testpaths = ["tests"]
pythonpath = ["src"]
asyncio_mode = "auto"

[tool.ruff]
line-length = 100
target-version = "py311"

[tool.ruff.lint]
select = [
    "E",   # pycodestyle errors
    "W",   # pycodestyle warnings
    "F",   # pyflakes
    "I",   # isort
    "B",   # flake8-bugbear
    "C4",  # flake8-comprehensions
    "UP",  # pyupgrade
]
ignore = [
    "E501",  # line too long (handled by formatter)
    "B008",  # do not perform function calls in argument defaults
]

[tool.ruff.format]
quote-style = "double"
indent-style = "space"

[tool.mypy]
python_version = "3.11"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = true
ignore_missing_imports = true
